mpsl: Add Bluetooth-compatible external radio coexistence interface
This adds the option of using a bluetooth-compatible external radio
coexistence system, using a 3-wire interface.

This is intended for 2.4GHz radios operating in very close proximity,
but not on the same SoC.

Note: as this implementation is not yet compatible with the previously
published API, it has been decided to separate the 802.15.4 and
Bluetooth implementations for the time being.

Signed-off-by: Jonathan Rico <jonathan.rico@nordicsemi.no>